+++ thold.php
@@ -1818,6 +1818,16 @@ function thold_edit() {
 				'default' => '',
 				'max_length' => '255',
 			),
+			'snmp_event_description' => array(
+				'friendly_name' => __('SNMP Event Description', 'thold'),
+				'description' => __('You can customize the event description being sent out to the SNMP notification receivers by using variable bindings. Standard Device (|host_*|), Data Query (|query_*|) substitution variables can be used as well as following varbinds:<br>&#060;THRESHOLDNAME&#062; &#060;HOSTNAME&#062; &#060;HOSTIP&#062; &#060;TEMPLATE_ID&#062; &#060;TEMPLATE_NAME&#062; &#060;THR_TYPE&#062; &#060;DS_NAME&#062; &#060;HI&#062; &#060;LOW&#062; &#060;EVENT_CATEGORY&#062; &#060;FAIL_COUNT&#062; &#060;FAIL_DURATION&#062;', 'thold'),
+				'method' => 'textarea',
+				'class' => 'textAreaNotes',
+				'textarea_rows' => '5',
+				'textarea_cols' => '80',
+				'value' => isset($thold_data['snmp_event_description']) ? $thold_data['snmp_event_description'] : '',
+				'default' => '',
+			),
 			'snmp_event_severity' => array(
 				'friendly_name' => __('SNMP Notification - Alert Event Severity', 'thold'),
 				'method' => 'drop_array',
@@ -2103,6 +2113,7 @@ function thold_edit() {
 		$('#trigger_cmd_norm').prop('disabled', status);
 
 		if ($('#snmp_event_category')) $('#snmp_event_category').prop('disabled', status);
+        if ($('#snmp_event_description')) $('#snmp_event_description').prop('disabled', status);
 		if ($('#snmp_event_severity')) $('#snmp_event_severity').prop('disabled', status);
 		if ($('#snmp_event_warning_severity')) $('#snmp_event_warning_severity').prop('disabled', status);