+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 * feature#469: use thold template name as filename for single export
 
+* feature#481: SNMP event description can be defined individually or template based. 
+
 --- 1.5.2 ---
 
 * issue#457: Page show blank if do export function without log info exist